backend/requirements.txt
fastapi
uvicorn
pydantic
groq
python-multipart
python-dotenv
firebase-admin
